refactor: app component
