+++ src/app-components/Carousel/CarouselStyle.css
